requirements.txt
flask
flask-sqlalchemy
flask-migrate
bcrypt
flask-mail
flask-jwt-extended
flask-marshmallow
marshmallow-sqlalchemy
psycopg2